# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ruckus Conference Program Operations API
  version: 1.0.0
  summary: Conference program operations - CFP, review, speakers, schedule, comms, CRM and public program feeds.
  description: >-
    HTTP API for Ruckus, an open-source conference program-operations tool. Paths are
    generated from the live Hono route registrations in src/app.ts and the src route
    modules, so path and method coverage is exact. Request and response bodies are
    intentionally loose object schemas with descriptions: handlers accept and return
    JSON envelopes shaped as a data property, and the demo does not enforce strict
    field level typing. Identity is demo persona simulation, not authentication.
  license:
    name: MIT
servers:
  - url: https://ruckus.to
    description: Live demo deployment
  - url: http://localhost:8787
    description: Local development API
...
security: []
paths:
...
  "/api/events/{eventId}/agenda/proposals":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    get:
      tags: [schedule-agenda]
      operationId: get_api_events_eventId_agenda_proposals
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/proposals/generate":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_proposals_generate
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/proposals/{id}/placements/{placementId}/{decision}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
      - name: placementId
        in: path
        required: true
        schema:
          type: string
        description: "placementId path parameter."
      - name: decision
        in: path
        required: true
        schema:
          type: string
        description: "decision path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_proposals_id_placements_placementId_decision
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/proposals/{id}/{decision}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: "id path parameter."
      - name: decision
        in: path
        required: true
        schema:
          type: string
        description: "decision path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_proposals_id_decision
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/publish":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_publish
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/rooms":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_rooms
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/rooms/{roomId}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: roomId
        in: path
        required: true
        schema:
          type: string
        description: "roomId path parameter."
    patch:
      tags: [schedule-agenda]
      operationId: patch_api_events_eventId_agenda_rooms_roomId
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/tracks":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    post:
      tags: [schedule-agenda]
      operationId: post_api_events_eventId_agenda_tracks
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/agenda/tracks/{trackId}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: trackId
        in: path
        required: true
        schema:
          type: string
        description: "trackId path parameter."
    patch:
      tags: [schedule-agenda]
      operationId: patch_api_events_eventId_agenda_tracks_trackId
      summary: "Registered in agendaRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  "/public/events/{eventId}/itinerary":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    get:
      tags: [public-widgets]
      operationId: get_public_events_eventId_itinerary
      summary: "Registered in publicSite.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
      description: >-
        Loosely typed JSON request body. Handlers read the fields they need and ignore
        the rest; see docs/API.md for the fields each operation consumes.
    DataEnvelope:
      type: object
      additionalProperties: true
      description: >-
        Standard success envelope. Most JSON endpoints return a data property holding an
        object or array; some also return meta describing counts or provider outcomes.
      properties:
        data:
          description: Operation payload, either an object or an array.
        meta:
          type: object
          additionalProperties: true
          description: Optional counts, filters or provider status details.
    ErrorEnvelope:
      type: object
      additionalProperties: true
      description: Standard failure envelope returned by the shared error helper.
      properties:
        error:
          type: object
          additionalProperties: true
          properties:
            code:
              type: string
              description: Short machine readable code such as NOT_FOUND or VALIDATION_ERROR.
            message:
              type: string
              description: Human readable explanation.